add AdjustableManager
The idea of the AdjustableManager is that if you want to change stuff
(PID params, current thresholds, etc.) during tuning, you either have to
adjust magic numbers and redeploy or write a bunch of annoying code to
update things. So, I wrote something that automates that–basically, if
you annotate a field as Tunable and pass the enclosing class to the
AdjustableManager, then it handles managing all the SmartDashboard stuff
for you. It also has a feature where you can annotate a field (currently
all WPILib I/O stuff) as Telemetry and it will likewise display its
output on the SmartDashboard. Untested but shouldn't need much testing
anyway.