Create SubsystemAttached interface
This interface gets around some of the problems with multiple
inheritance by requiring all sub-implementors to provide a Subsystem
instance. This will produce some rather unsightly boilerplate but it's
better than nothing. We could even use a default method that checks if
the implementor is an instance of Subsystem and throws an exception if
it isn't.