Create VibrationManager
The VibrationManager, as it says in the docs, is used for managing
joystick vibration. It's built around the command-subsystem framework,
because it turns out that TimedCommands are the perfect system if you
want to run some initialization code (namely setRumble(rumbleIntensity))
, wait a given amount of time, and then run some de-initialization code
(setRumble(0)). This also has the benefit of possibly making it more
feature-rich later (this is explained in the javadoc for
JoystickSubsystem.)

This is in the subsystem-templates branch because it's necessary for
properly implementing the PID actuator feature that I'm porting over
from Kingbass's code. On Kingbass there's a thing that makes the
controller buzz when the arm hits a limit, so I'm planning to implement
that here as well.

All the code that does things here is a direct
copy-paste from Kingbass; I've just added some code to make it all
self-contained (just keeping a map of JoystickSubsystems by
joystick ID so that requires() checks work properly with arbitrary
joysticks.)