Improve throttle-limiting logic
Now doesn't limit the throttle for one direction arbitrarily when motors
are stopped.

Before this, if the bot had just decelerated from forwards motion,
the backward throttle would have been limited until the robot cleared
the "deadzone". Now takes into account whether or not the robot is
moving.

May need further refinements based on joystick control so that if we are
in a pushing match the motors won't be limited if we start getting
pushed backwards (could be a problem with non-traction-limited
drivetrains).