Create new deadzone function
Keeping the old one around because we respect backwards compatibility.

Because I wanted fancy equations, this also adds a MathJax renderer to
the generated Javadoc files.

Since the fancy TeX Javadoc doesn't render in IDE's, we may
want to find a solution such as deploying a copy of the javadoc to
a Github Page with Travis.

Signed-off-by: Zachary Robinson <robinsonz@catlin.edu>