Fix bug with ceilings not being respected properly
Basically if ceilings were updated while the power management wasn't scaling, they weren't respected. That has been fixed by checking if the power management is scaling, and if not, manually updating the peakOutputForwards and reverse. I'm not super happy with this solution as it massively complicates things and makes it that a single value is being updated in three places rather than one, but it's really the only reasonable solution I can think of.