Make robot class property work nicely
Now the jar task checks if the deploy task will run; if so, it does the
checking if the robot class is set. This is a good deal better than the
earlier band-aid solution which was to set it to a default.
Now the build will succeed if robotClass is not set but we aren't
deploying, but will fail if it is not set and we are deploying.