Set slave motors to 0 on deinitialization
When this wasn't happening (i.e. we just set motors 2 or 3 to null) then
the motor instance would get garbage collected (I think, it could be
referenced somewhere else), but the motor itself would still obey the
previous command which was to follow the first motors. So if you changed
a motor ID from something to -1 then the motor which you thought that
you weren't commanding anymore actually would still be running.