Update to 2019
Whew. So, the 2019 GradleRIO and WPILib update borked a lot of the
library structure (fortunately almost none of the code) and so I've
restructured the thing. Again. Now, instead of having two gradle
projects, we have two source sets (i.e. folders in src/). There's still
some improvements that need to be made.

 * The choose-your-own-robot-class function is entirely broken and I'm
   not sure how I can fix it without some ugly stuff.

 * Not totally sure about new API stuff, and various bits and bobs may
   need to be migrated for them to work properly.

 * The testing framework is temporarily deleted because CTRE updates
   borked it. @edelmanjm