Re-add run configuration robot class deployment
WPILib doesn’t specify the robot class in the manifest anymore (you
instead define a main function that instantiates your robot of choice),
so the run-configuration-robot-class-thing broke. This is a really ugly
fix.

According to WPILib, you aren't supposed to modify the main function at
all. Therefore, I thought it prudent to add roughly 70 lines of code
doing all sorts of interesting filesystem and reflection things.
Basically, now what this does is the custom createRobotclassTxt task
reads the robotClass project property (which used to be passed into
GradleRIO) and writes it to a text file (specifically,
build/robotclass.txt). GradleRIO then deploys robotclass.txt to the
robot, where TestbotLoaderMain (which is the "main class" for all
ROOSTER test robots) reads it and attempts to instantiate the class
named within. Now, while I concede this is a terrible system, it is only
marginally worse than what WPILib/GradleRIO used to be doing (namely,
putting the robot class in the jar manifest and reading that.) So I
think that I'm just being ingenious.