Link to Java 11 API docs
This *should* fix the build errors. However, there seems to be a bug in
the JDK where this doesn't actually link properly, so we're just gonna
roll with that. I'm leaving it in here because it keeps the docs looking
nice and it's not like anyone is going to be browsing our docs and
suddenly go "wait, I forgot what a String is".

See: https://bugs.openjdk.java.net/browse/JDK-8211194